+++ src/service.rs
@@ -1,92 +1,132 @@
-use anyhow::Error;
-use rtw::{
-    Activity, ActivityId, ActivityService, CurrentActivityRepository, DateTimeW,
-    FinishedActivityRepository, OngoingActivity,
-};
+use crate::rtw_core::activity::{intersect, overlap, Activity, OngoingActivity};
+use crate::rtw_core::datetimew::DateTimeW;
+use crate::rtw_core::service::ActivityService;
+use crate::rtw_core::storage::Storage;
+use crate::rtw_core::ActivityId;
+use anyhow::anyhow;
 
-pub struct Service<F, C>
+pub struct Service<S>
 where
-    F: FinishedActivityRepository,
-    C: CurrentActivityRepository,
+    S: Storage,
 {
-    finished: F,
-    current: C,
+    storage: S,
 }
 
-impl<F, C> Service<F, C>
+impl<S> Service<S>
 where
-    F: FinishedActivityRepository,
-    C: CurrentActivityRepository,
+    S: Storage,
 {
-    pub fn new(finished: F, current: C) -> Self {
-        Service { finished, current }
+    pub fn new(storage: S) -> Self {
+        Service { storage }
     }
 }
 
-impl<F, C> ActivityService for Service<F, C>
+impl<S> ActivityService for Service<S>
 where
-    F: FinishedActivityRepository,
-    C: CurrentActivityRepository,
+    S: Storage,
 {
     fn get_current_activity(&self) -> anyhow::Result<Option<OngoingActivity>> {
-        self.current.get_current_activity()
+        self.storage.get_current_activity().map_err(|e| e.into())
     }
 
     fn start_activity(&mut self, activity: OngoingActivity) -> anyhow::Result<OngoingActivity> {
-        self.stop_current_activity(activity.start_time)?;
-        let started = OngoingActivity::new(activity.start_time, activity.tags);
-        self.current.set_current_activity(started.clone())?;
-        Ok(started)
+        let finished = self.storage.get_finished_activities()?;
+        let intersections = time_intersections(finished.as_slice(), &activity.start_time);
+        if intersections.is_empty() {
+            self.stop_current_activity(activity.start_time)?;
+            self.storage.set_current_activity(activity.clone())?;
+            Ok(activity)
+        } else {
+            Err(anyhow!("{:?} would overlap {:?}", activity, intersections))
+        }
     }
 
     fn stop_current_activity(&mut self, time: DateTimeW) -> anyhow::Result<Option<Activity>> {
-        let current = self.current.get_current_activity()?;
+        let current = self.storage.get_current_activity()?;
         match current {
             None => Ok(None),
             Some(current_activity) => {
-                self.finished
-                    .write_activity(current_activity.clone().into_activity(time)?)?;
-                self.current.reset_current_activity()?;
-                Ok(Some(current_activity.into_activity(time)?))
+                let stopped = current_activity.clone().into_activity(time)?;
+                let finished = self.storage.get_finished_activities()?;
+                let intersections = activity_intersections(finished.as_slice(), &stopped);
+                if intersections.is_empty() {
+                    self.storage.write_activity(stopped)?;
+                    self.storage.reset_current_activity()?;
+                    Ok(Some(current_activity.into_activity(time)?))
+                } else {
+                    Err(anyhow!("{:?} would overlap {:?}", stopped, intersections))
+                }
             }
         }
     }
 
-    fn filter_activities<P>(&self, p: P) -> Result<Vec<(ActivityId, Activity)>, Error>
+    fn cancel_current_activity(&mut self) -> anyhow::Result<Option<OngoingActivity>> {
+        self.storage.reset_current_activity().map_err(|e| e.into())
+    }
+
+    fn filter_activities<P>(&self, p: P) -> anyhow::Result<Vec<(ActivityId, Activity)>>
     where
         P: Fn(&(ActivityId, Activity)) -> bool,
     {
-        self.finished.filter_activities(p)
+        self.storage.filter_activities(p).map_err(|e| e.into())
     }
 
-    fn delete_activity(&self, id: ActivityId) -> Result<Option<Activity>, Error> {
-        self.finished.delete_activity(id)
+    fn get_finished_activities(&self) -> anyhow::Result<Vec<(ActivityId, Activity)>> {
+        self.storage.get_finished_activities().map_err(|e| e.into())
     }
 
-    fn track_activity(&mut self, activity: Activity) -> Result<Activity, Error> {
-        self.finished.write_activity(activity.clone())?;
-        Ok(activity)
+    fn delete_activity(&self, id: ActivityId) -> anyhow::Result<Option<Activity>> {
+        self.storage.delete_activity(id).map_err(|e| e.into())
     }
+
+    fn track_activity(&mut self, activity: Activity) -> anyhow::Result<Activity> {
+        let finished = self.storage.get_finished_activities()?;
+        let intersections = activity_intersections(finished.as_slice(), &activity);
+        if intersections.is_empty() {
+            self.storage.write_activity(activity.clone())?;
+            Ok(activity)
+        } else {
+            Err(anyhow!("{:?} would overlap {:?}", activity, intersections))
+        }
+    }
+}
+
+fn activity_intersections(
+    activities: &[(ActivityId, Activity)],
+    activity: &Activity,
+) -> Vec<Activity> {
+    activities
+        .iter()
+        .filter_map(|(_, a)| overlap(a, activity))
+        .collect()
+}
+
+fn time_intersections(
+    activities: &[(ActivityId, Activity)],
+    start_time: &DateTimeW,
+) -> Vec<Activity> {
+    activities
+        .iter()
+        .filter_map(|(_, a)| intersect(a, start_time))
+        .collect()
 }
 
 #[cfg(test)]
 mod tests {
     use crate::chrono_clock::ChronoClock;
-    use crate::json_current::JsonCurrentActivityRepository;
-    use crate::json_finished::JsonFinishedActivityRepository;
+    use crate::json_storage::JsonStorage;
+    use crate::rtw_core::activity::OngoingActivity;
+    use crate::rtw_core::clock::Clock;
+    use crate::rtw_core::datetimew::DateTimeW;
+    use crate::rtw_core::service::ActivityService;
     use crate::service::Service;
-    use rtw::{ActivityService, Clock, DateTimeW, OngoingActivity};
+    use chrono::{Local, TimeZone};
     use tempfile::{tempdir, TempDir};
 
-    fn build_json_service(
-        test_dir: &TempDir,
-    ) -> Service<JsonFinishedActivityRepository, JsonCurrentActivityRepository> {
-        let writer_path = test_dir.path().join(".rtww.json");
-        let repository_path = test_dir.path().join(".rtwr.json");
-        Service::new(
-            JsonFinishedActivityRepository::new(writer_path),
-            JsonCurrentActivityRepository::new(repository_path),
-        )
+    fn build_json_service(test_dir: &TempDir) -> Service<JsonStorage> {
+        let finished_path = test_dir.path().join(".rtwh.json");
+        let current_path = test_dir.path().join(".rtwc.json");
+        Service::new(JsonStorage::new(current_path, finished_path))
     }
 
     #[test]
@@ -152,6 +192,75 @@ mod tests {
     }
 
     #[test]
+    fn test_start_intersecting_activity() {
+        let test_dir = tempdir().expect("error while creating tempdir");
+        let mut service = build_json_service(&test_dir);
+        let finished = OngoingActivity::new(
+            Local
+                .datetime_from_str("2020-12-25T09:00:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+            vec![],
+        )
+        .into_activity(
+            Local
+                .datetime_from_str("2020-12-25T10:00:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+        )
+        .unwrap();
+        let tracked = service.track_activity(finished);
+        assert!(tracked.is_ok());
+        let other = OngoingActivity::new(
+            Local
+                .datetime_from_str("2020-12-25T09:30:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+            vec![],
+        );
+        let started = service.start_activity(other);
+        assert!(started.is_err());
+    }
+
+    #[test]
+    fn test_stop_intersecting_activity() {
+        let test_dir = tempdir().expect("error while creating tempdir");
+        let mut service = build_json_service(&test_dir);
+        let finished = OngoingActivity::new(
+            Local
+                .datetime_from_str("2020-12-25T09:00:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+            vec![],
+        )
+        .into_activity(
+            Local
+                .datetime_from_str("2020-12-25T10:00:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+        )
+        .unwrap();
+        let tracked = service.track_activity(finished);
+        assert!(tracked.is_ok());
+        let other = OngoingActivity::new(
+            Local
+                .datetime_from_str("2020-12-25T08:30:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+            vec![],
+        );
+        let started = service.start_activity(other);
+        assert!(started.is_ok());
+        let stopped = service.stop_current_activity(
+            Local
+                .datetime_from_str("2020-12-25T09:30:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+        );
+        assert!(stopped.is_err());
+    }
+
+    #[test]
     fn test_summary_nothing() {
         let clock = ChronoClock {};
         let test_dir = tempdir().expect("error while creating tempdir");
@@ -201,5 +310,43 @@ mod tests {
             range_start <= a.get_start_time() && a.get_start_time() <= range_end
         });
         assert!(activities.unwrap().is_empty());
+    }
+
+    #[test]
+    fn test_track_intersecting_activity() {
+        let test_dir = tempdir().expect("error while creating tempdir");
+        let mut service = build_json_service(&test_dir);
+        let finished = OngoingActivity::new(
+            Local
+                .datetime_from_str("2020-12-25T09:00:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+            vec![],
+        )
+        .into_activity(
+            Local
+                .datetime_from_str("2020-12-25T10:00:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+        )
+        .unwrap();
+        let tracked = service.track_activity(finished);
+        assert!(tracked.is_ok());
+        let other = OngoingActivity::new(
+            Local
+                .datetime_from_str("2020-12-25T09:30:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+            vec![],
+        )
+        .into_activity(
+            Local
+                .datetime_from_str("2020-12-25T10:30:00", "%Y-%m-%dT%H:%M:%S")
+                .unwrap()
+                .into(),
+        )
+        .unwrap();
+        let tracked = service.track_activity(other);
+        assert!(tracked.is_err());
     }
 }