+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0](https://crates.io/crates/rtw/1.2.0) Jun 07, 2020
+
+* add `cancel` subcommand.
+* deny overlapping activities
+* add `timeline` subcommand.
+* timeline colors can be configured in `rtw_config.json`
+* add `day` subcommand (display timeline for the current day)
+* add `week` subcommand (display timeline for the current week)
+
 ## [1.1.0](https://crates.io/crates/rtw/1.1.0) Mar 22, 2020
 
 ### Added